+++ app/Http/Controllers/Admin/ConstituenciesController.php
@@ -6,12 +6,15 @@ use App\County;
 use App\Constituency;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\UpsertConstituencyRequest;
 
 class ConstituenciesController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view_constituencies_page');
+        
         $constituencies = Constituency::with(['county', 'wards'])->get();
 
         return view('admin.constituencies.index', compact('constituencies'));
@@ -19,6 +22,8 @@ class ConstituenciesController extends Controller
 
     public function create()
     {
+        Gate::authorize('view_create_constituency_page');
+
         $counties = County::all();
 
         return view('admin.constituencies.create', compact('counties'));
@@ -33,11 +38,14 @@ class ConstituenciesController extends Controller
 
     public function show(Constituency $constituency)
     {
+        Gate::authorize('view_single_constituency_page');
         return view('admin.constituencies.show', compact('constituency'));
     }
 
     public function edit(Constituency $constituency)
     {
+        Gate::authorize('view_edit_constituency_page');
+
         $counties = County::all();
 
         return view('admin.constituencies.edit', compact('constituency', 'counties'));
@@ -46,12 +54,16 @@ class ConstituenciesController extends Controller
 
     public function update(UpsertConstituencyRequest $request, Constituency $constituency)
     {
+        Gate::authorize('update_constituency');
+
         $constituency->update($request->validated());
 
         return redirect()->route('admin.constituencies.show', $constituency);
     }
 
     public function destroy(Constituency $constituency){
+        
+        Gate::authorize('delete_constituency');
 
         $constituency->delete();
