+++ app/Http/Controllers/Admin/OfficersController.php
@@ -6,6 +6,7 @@ use App\Role;
 use App\User;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 
 class OfficersController extends Controller
 {
@@ -16,6 +17,8 @@ class OfficersController extends Controller
      */
     public function index()
     {
+        Gate::authorize('view_officers_page');
+
         $users = User::with('working.station')->role(Role::findOrFail(3))->get();
 
         return view('admin.officers.index', compact('users'));
@@ -30,6 +33,8 @@ class OfficersController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('delete_officer');
+        
         $user = User::findOrFail($id);
 
         if(!is_null($user->working)) $user->working->delete();
