+++ app/Http/Controllers/Admin/WardsController.php
@@ -6,6 +6,7 @@ use App\Ward;
 use App\Constituency;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\UpsertWardRequest;
 
 class WardsController extends Controller
@@ -13,6 +14,8 @@ class WardsController extends Controller
     
     public function index()
     {
+        Gate::authorize('view_wards_page');
+        
         $wards = Ward::with(['constituency.county', 'users'])->get();
         
         return view('admin.wards.index', compact('wards'));
@@ -20,6 +23,8 @@ class WardsController extends Controller
 
     public function create()
     {
+        Gate::authorize('view_create_ward_page');
+
         $constituencies = Constituency::all();
 
         return view('admin.wards.create', compact('constituencies'));
@@ -27,10 +32,14 @@ class WardsController extends Controller
 
     public function show(Ward $ward)
     {
+        Gate::authorize('view_single_ward_page');
+
         return view('admin.wards.show', compact('ward'));
     }
 
     public function store(UpsertWardRequest $request){
+
+        Gate::authorize('store_ward');
 
         Ward::create($request->validated());
 
@@ -39,6 +48,8 @@ class WardsController extends Controller
 
     public function edit(Ward $ward)
     {
+        Gate::authorize('view_edit_ward_page');
+
         $constituencies = Constituency::all();
 
         return view('admin.wards.edit', compact('ward', 'constituencies'));
@@ -47,6 +58,8 @@ class WardsController extends Controller
 
     public function update(UpsertWardRequest $request, Ward $ward)
     {
+        Gate::authorize('update_ward');
+
         $ward->update($request->validated());
 
         return redirect()->route('admin.wards.show', $ward);
@@ -54,6 +67,8 @@ class WardsController extends Controller
 
     public function destroy(Ward $ward)
     {
+        Gate::authorize('delete_ward');
+        
         $ward->delete();
 
         return redirect()->route('admin.wards.index');
